requirements.txt
streamlit>=1.31.0
chromadb>=0.4.22
sentence-transformers>=2.5.0
numpy>=1.24.0,<2.0.0
openai>=1.12.0
anthropic>=0.18.0
google-generativeai>=0.4.0
ollama>=0.1.6
python-dotenv>=1.0.0
requests>=2.31.0
